fix: duplicate code in datasource providers
